+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   * Support model experimentl access levels when creating or editng projects.
   * Support `unidiff` requests for merge request diffs.
   * Support the `SecurityOrchestrationPolicy` for the pipeline source filter.
+  * Support returning users marked as auditors when listing all users.
 
 ## Deprecations